Reformed charge code to make it more stable and public
Moved charge subclasses out

Renamed charge concepts to make them easier to understand

Fixed force track emitter not sending state updates

Moved tile interfaces

Fixed #1413 redstone interaction of admin charge feeders

+++ src/main/java/mods/railcraft/common/blocks/tracks/behaivor/HighSpeedTools.java
@@ -94,8 +94,8 @@ public class HighSpeedTools {
         return TrackTools.getTrackTypeAt(world, pos).isHighSpeed();
     }
 
-    public static float speedForNextTrack(World world, BlockPos pos, int dist, @Nullable EntityMinecart cart) {
-        float maxSpeed = RailcraftConfig.getMaxHighSpeed();
+    public static double speedForNextTrack(World world, BlockPos pos, int dist, @Nullable EntityMinecart cart) {
+        double maxSpeed = RailcraftConfig.getMaxHighSpeed();
         if (dist < LOOK_AHEAD_DIST)
             for (EnumFacing side : EnumFacing.HORIZONTALS) {
                 BlockPos nextPos = pos.offset(side);
